+++ CHANGES.md
@@ -6,6 +6,7 @@ I did not browse all issues on github at first, so I did not recognize that some
 
 ## History
 
+* Feature: In Yaml, Advances allows -1 as an indicator to represent "all advances", e.g. for debugging or testing purposes. The mapper will resolve this to the full list of advance IDs in the game.
 * Feature: Gameplay updates
   * Future Tech handling:
     * Added per-player `FutureTechCount` runtime state.